% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adap_samp.R
\name{adap_samp}
\alias{adap_samp}
\title{Adaptive Envelope Rejection Sampling for Log-Concave Densities}
\usage{
adap_samp(n, density, density_deriv, p, seed = NULL, zb = c(-Inf, Inf))
}
\description{
Adaptive Envelope Rejection Sampling for Log-Concave Densities
}
