% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_val_year.R
\name{cs_checkCodedMonth}
\alias{cs_checkCodedMonth}
\title{Check Month for Single Coded Month Value}
\usage{
cs_checkCodedMonth(monthItem)
}
\arguments{
\item{monthItem}{A single item in a year list object}
}
\description{
Ensure that there is only one coded month value per year list object item
}
\keyword{internal}
