% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_val_year.R
\name{cs_checkMonth}
\alias{cs_checkMonth}
\title{Check Month of Year List Object for Correct Properties}
\usage{
cs_checkMonth(monthItem)
}
\arguments{
\item{monthItem}{A single item in a year list object}
}
\description{
This checks a single month for the correct properties
}
\keyword{internal}
