% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_val_year.R
\name{cs_checkNames}
\alias{cs_checkNames}
\title{Validate Number of Months}
\usage{
cs_checkNames(.data)
}
\arguments{
\item{.data}{A yaer list object}
}
\description{
This checks the list of months to make sure that there are either
    12 unique month entries or, if there are fewer than 12, that there are no
    missing missing months between January and the last month loaded.
}
\details{
For year list objects with fewer than 12 months, this confirms that
    data for January are present. It then ensures that there are no months
    between January and the last month entered. For example, if the first six
    months of data were loaded - i.e. through June - this function would
    ensure that there were no months missing between January and June.
}
\keyword{internal}
