% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validate_year.R
\name{cs_checkVarClasses}
\alias{cs_checkVarClasses}
\title{Check Variable Classes}
\usage{
cs_checkVarClasses(monthItem)
}
\arguments{
\item{monthItem}{A single item in a year list object}
}
\description{
Checks to make sure variable classes are correct
}
\keyword{internal}
