% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validate_year.R
\name{cs_checkVarNames}
\alias{cs_checkVarNames}
\title{Check Variable Names}
\usage{
cs_checkVarNames(monthItem)
}
\arguments{
\item{monthItem}{A single item in a year list object}
}
\description{
Checks to make sure variable names are correct if there are 20 variables in month
}
\keyword{internal}
