% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validate_year.R
\name{cs_cleanResults}
\alias{cs_cleanResults}
\title{Clean Results Tibble}
\usage{
cs_cleanResults(.data)
}
\arguments{
\item{.data}{A result tibble}
}
\description{
Converts messy results output to clean output
}
\keyword{internal}
