% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_collapse.R
\name{cs_collapse}
\alias{cs_collapse}
\title{Combine Months in Year List Object}
\usage{
cs_collapse(.data)
}
\arguments{
\item{.data}{A list containing monthly crime data}
}
\value{
A tibble containing a year worth of crime data.
}
\description{
This function takes a list containing 12 tibbles - one per
   month - that have been validated and collapses them into a single tibble.
   This is preferable for working with the data post validation, but cannot
   be done until there is parity in terms of variable names and classes.
}
