% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_combine.R
\name{cs_combine}
\alias{cs_combine}
\title{Combine Multiple Years of Data}
\usage{
cs_combine(type = "year", date, ...)
}
\arguments{
\item{type}{"year" is the only valid input currenty; year to date functionality
is planned for a later update}

\item{date}{For \code{type = "year"}, this should be the year of data to be returned.
For \code{type = "ytd"}, this should be the last month to be included in each estimate.}

\item{...}{An unquoted list of objects}
}
\value{
A tibble containing a selection of combined crime data for a given time period.
}
\description{
This function...
}
