% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{cs_create_index}
\alias{cs_create_index}
\title{Create Index of Available Months}
\usage{
cs_create_index()
}
\value{
A tibble with all available monthly crime tables, the \code{iframe} page
    they appear on, and their row number.
}
\description{
Constructs a table for finding a given table of crime data or a set of
    tables (such as year to date or full year). This is largely needed for internal
    use when downloading tables, but is exported for reference and troubleshooting.
}
\examples{
# create index
\donttest{
i <- cs_create_index()

# preview of index object
i
}

}
