% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_crime_filter.R
\name{cs_crime_filter}
\alias{cs_crime_filter}
\title{Filter Crimes}
\usage{
cs_crime_filter(.data, var, crime)
}
\arguments{
\item{.data}{A tbl}

\item{var}{Name of variable with 5 or 6 digit crime codes}

\item{crime}{A string describing the crime type to be identified}
}
\value{
A subset tibble with only the specified crimes
}
\description{
\code{cs_crime_filter} can be used to subset based on
    specific single UCR categories or common groupings.
}
