% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_filter_count.R
\name{cs_filter_count}
\alias{cs_filter_count}
\title{Remove Negative Counts}
\usage{
cs_filter_count(.data, variable)
}
\arguments{
\item{.data}{A tbl}

\item{variable}{the name of the column}
}
\value{
returns the data frame with the rows containing -1 removed
}
\description{
Removes the row that contains -1 in a specified column
}
