% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_filter_crime.R
\name{cs_filter_crime}
\alias{cs_filter_crime}
\title{Filter Crimes}
\usage{
cs_filter_crime(.data, var, crime)
}
\arguments{
\item{.data}{A tbl}

\item{var}{Name of variable with 5 or 6 digit crime codes}

\item{crime}{A string describing the crime type to be identified}
}
\value{
A subset tibble with only the specified crimes
}
\description{
\code{cs_filter_crime} can be used to subset based on
    specific single UCR categories or common groupings.
}
\details{
The categories used here are derived from the U.S. Federal
    Bureau of Investigation's Uniform Crime Reporting codes.
}
\examples{
testData <- january2018
testData <- cs_filter_crime(testData,Crime,"violent")

}
