% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{cs_get_data}
\alias{cs_get_data}
\title{Download Crime Data from SLMPD}
\usage{
cs_get_data(year, month)
}
\arguments{
\item{year}{A year value in the style \code{YYYY}}

\item{month}{A month number}
}
\value{
A year-list object ready for validation.
}
\description{
Downloads crime data from the SLMPD website.
}
