% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_load_year.R
\name{cs_identifyMonth}
\alias{cs_identifyMonth}
\title{Extract Month of a Given Year List Object Item}
\usage{
cs_identifyMonth(.data, read = TRUE)
}
\arguments{
\item{.data}{A year list object name}

\item{read}{A logical scalar; if \code{TRUE}, return output structured for initial reading of
data into R. If \code{FALSE}, use for data validation.}
}
\description{
This uses the value of the first observation's coded month as the basis for
    identifying which month the data are from.
}
\keyword{internal}
