% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_isIntersection.R
\name{cs_isIntersection}
\alias{cs_isIntersection}
\title{Is A Location An Intersection?}
\usage{
cs_isIntersection(.data, variable)
}
\arguments{
\item{.data}{A tbl}

\item{variable}{A column containing \code{ILEADSAddress} data}
}
\value{
a logical vector that displays \code{TRUE} where the column is a
    intersection and displays \code{FALSE} when the column isn't an
    intersection.
}
\description{
The \code{ILEADSAddress} field contains intersections as well
    as incidents located at a single street address. This function can be
    used to identify intersections for specific operations.
}
