% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_load_year.R
\name{cs_load_year}
\alias{cs_load_year}
\title{Create Year List Object}
\usage{
cs_load_year(path)
}
\arguments{
\item{path}{A file path}
}
\value{
A year list object containing 12 tibbles - one per month - worth
   of crime data.
}
\description{
\code{cs_load_year} is used to load a set of \code{.csv} files
   contained in the given directory. This should be used to load a full
   year worth of data.
}
