% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_load_year.R
\name{cs_matchMonth}
\alias{cs_matchMonth}
\title{Match Extract Month with Month Name}
\usage{
cs_matchMonth(x)
}
\arguments{
\item{x}{The last two characters of the first observation's coded month value}
}
\description{
Based on the result of cs_identifyMonth, this function returns the
    appropriate string name.
}
\keyword{internal}
