% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_val_year.R
\name{cs_matchMonthNum}
\alias{cs_matchMonthNum}
\title{Match List Name with Numeric Value}
\usage{
cs_matchMonthNum(x)
}
\arguments{
\item{x}{The last two characters of the first observation's coded month value}
}
\description{
.
}
\keyword{internal}
