% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_val_year.R
\name{cs_matchMonths}
\alias{cs_matchMonths}
\title{Logic Check for Named and Coded Month}
\usage{
cs_matchMonths(.data)
}
\arguments{
\item{.data}{A result tibble}
}
\description{
Updates test result tibble with results of logic check that compares the
    data from the named month with the data from the coded month.
}
\keyword{internal}
