% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validate_year.R
\name{cs_matchYear}
\alias{cs_matchYear}
\title{Logic Check for Named and Coded Month}
\usage{
cs_matchYear(.data, year)
}
\arguments{
\item{.data}{A result tibble}
}
\description{
Updates test result tibble with results of logic check that compares the
    data from the given year argument with the data from the coded year.
}
\keyword{internal}
