% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_missing_xy.R
\name{cs_missing_xy}
\alias{cs_missing_xy}
\title{Missing Coordinates}
\usage{
cs_missing_xy(.data, varx, vary, newVar)
}
\arguments{
\item{.data}{a data frame}

\item{varx}{the column from the data frame containing the x coordinates}

\item{vary}{the column from the data frame containg the y coordinates}

\item{newVar}{the name of the new column}
}
\value{
returns a logical vector that displays TRUE when the coordinates are
    missing and FALSE if the coordinates are not missing
}
\description{
\code{cs_missingCoords} compares X and Y coordinates and adds a
    TRUE/FALSE column that states which rows are missing coordinate
    data.
}
\examples{
cs_filter_crime(january2018,XCoord,Ycoord,IsMissing)

}
