% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_parse_int.R
\name{cs_parse_int}
\alias{cs_parse_int}
\title{Seperate Interection}
\usage{
cs_parse_int(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{the column containing street adresses}
}
\value{
returns two columns containing street addresses and removes the input column
}
\description{
Takes a column that contains street addresses and seperates
    the addresses into two columns
}
