% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_parse_month.R
\name{cs_parse_month}
\alias{cs_parse_month}
\title{Seperate Coded Month}
\usage{
cs_parse_month(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{the variable containing coded month and coded year}
}
\value{
returns the data frame with two new columns named "codedYear" and "codedMonth" and the input column removed
}
\description{
Seperates a column containing coded year and coded month
    seperated by "-" into two columns and removes the input column
}
