% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_prep_year.R
\name{cs_prep_year}
\alias{cs_prep_year}
\title{Prepare Raw Data}
\usage{
cs_prep_year(path)
}
\arguments{
\item{path}{File path where raw STLMPD data are located}
}
\value{
A list containing old file names and new file names for reference.
}
\description{
Data downloaded from the St. Louis Metropolitan Police Department
   are downloaded with incorrect file paths - e.g. \code{January2008.CSV.html}.
   This function iterates over all files in a given path and replaces their
   file extensions. Thus \code{January2008.CSV.html} will be replaced by
   \code{january2008.csv}.
}
\examples{
\dontrun{
    cs_prep_year(path = "data/raw/2008")
}

}
