% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_project.R
\name{cs_project}
\alias{cs_project}
\title{Project Data}
\usage{
cs_project(.data, varX, varY, crs)
}
\arguments{
\item{.data}{A tibble with crime data}

\item{varX}{Name of column containing x coordinate data}

\item{varY}{Name of column containing y coordinate data}

\item{crs}{integer with the EPSG code, or character with proj4string representing the
coordinate reference system}
}
\description{
\code{cs_project} converts STLMPD data into a simple
    features object using the \code{XCoord} and \code{YCoord} columns.
}
