% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{cs_projectXY}
\alias{cs_projectXY}
\title{Project Data}
\usage{
cs_projectXY(.data, varX, varY, crs)
}
\arguments{
\item{.data}{A tibble with crime data}

\item{varX}{Name of column containing x coordinate data}

\item{varY}{Name of column containing y coordinate data}

\item{crs}{integer with the EPSG code, or character with proj4string representing the
coordinate reference system}
}
\value{
A \code{sf} object with the crime data projected for mapping.
}
\description{
\code{cs_projectXY} converts STLMPD data into a simple
    features object using the \code{XCoord} and \code{YCoord} columns.
}
