% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_replace_na.R
\name{cs_replace_na}
\alias{cs_replace_na}
\title{Replace Zeros with NA}
\usage{
cs_replace_na(.data, var)
}
\arguments{
\item{.data}{A tbl}

\item{var}{a column from the data frame}
}
\value{
returns the data frame with the 0's in the specified column changed
}
\description{
This function a specified column from the data frame and
    replaces cells that ahve the value 0 with NA
}
\examples{
testData <- january2018
testData <- cs_replace_na(testData, XCoord)
testData <- cs_replace_na(testData, YCoord)

}
