% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validate.R
\name{cs_validate}
\alias{cs_validate}
\title{Validate Year List Object}
\usage{
cs_validate(.data, year, verbose)
}
\arguments{
\item{.data}{A tbl}

\item{year}{A string representing the year being checked, e.g. \code{"2008" }}

\item{verbose}{A logical scalar. If \code{TRUE}, a full validation report summarizing
results will be returned. If \code{FALSE}, a single value will be returned.}
}
\value{
A tibble with validation results.
}
\description{
Data from SLMPD are released with a number of problems that
    \code{cs_validate_year} is designed to identify.
}
\details{
Since the data are imported alphabetically, \code{cs_validate_year}
    checks to ensure that items are imported in the correct order. It also
    checks each month to make sure it contains only one month worth of data.
    After ensuring that months are being read correctly, the function also
    counts the number of variables in each object. For all months prior to
    2013 and approximately half of the months during 2013, SLMPD data are
    released with 18 variables. For one month, May 2017, the data are released
    with 26 variables.
}
\examples{
\dontrun{
yearList08 <- cs_load_year(path = "data/raw/2008")

cs_validate(yearList08)
cs_validate(yearList08, verbose = TRUE)
}

}
