% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validate_month.R
\name{cs_validate_month}
\alias{cs_validate_month}
\title{Validate Variable Classes}
\usage{
cs_validate_month(.data, month)
}
\arguments{
\item{.data}{A tbl}

\item{month}{An option string name or abbreviation of a month, or its numeric value.
Acceptable inputs include, for example, "January", "january", "Jan",
"jan", and 1.}
}
\value{

}
\description{
This function can be applied to a single month within a year
    list object to identify problematic variables that are non-standard classes.
}
\details{
When the data are read into R, they occasionally are read in in ways
    that \code{readr}'s parser incorrectly handles. This issues can be identified
    using \code{cs_validate_year} as occuring in specific months.
    \code{cs_validate_month} is meant to be a follow-up function that gives more
    detail about specific variables that need to be addressed.
}
\seealso{
cs_validate_year
}
