% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validate_year.R
\name{cs_validateNames}
\alias{cs_validateNames}
\title{Validate Names of Months}
\usage{
cs_validateNames(names)
}
\arguments{
\item{names}{Vector of names of months from year list object}
}
\description{
Given the names of months in the year list object,
    are all months that should be present included?
}
\keyword{internal}
