% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfit_t.R
\name{emfit_t}
\alias{emfit_t}
\title{EM Implementation for Fitting Generalized t-Distribution}
\usage{
emfit_t(x, nu, maxit = 500L, min.eps = 1e-07, par = NULL)
}
\description{
EM Implementation for Fitting Generalized t-Distribution
}
