% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_isIntersection.R
\name{isIntersection}
\alias{isIntersection}
\title{Is Intersection}
\usage{
isIntersection(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{the column containing ILEADSAddress}
}
\value{
a logical vector that displays "TRUE" where the column is a intersection and displays "FALSE" when the column isn't an intersection
}
\description{
Is Intersection
}
