% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_missingCoords.R
\name{missingCoords}
\alias{missingCoords}
\title{missingCoords compares X and Y coordinates and adds a TRUE/FALSE column that states which rows are missing coordinate data}
\usage{
missingCoords(.data, variable1, variable2)
}
\arguments{
\item{.data}{a data frame}

\item{variable1}{the column from the data frame containing the x coordinates}

\item{variable2}{the column from the data frame containg the y coordinates}
}
\value{
returns a logical vector that displays TRUE when the coordinates are missing and FALSE if the coordinates are not missing
}
\description{
missingCoords compares X and Y coordinates and adds a TRUE/FALSE column that states which rows are missing coordinate data
}
