% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_raphson.R
\name{newton_raphson}
\alias{newton_raphson}
\title{Newton-Raphson Optimization}
\usage{
newton_raphson(
  par,
  H,
  gr,
  hess,
  d = 0.8,
  c = 0.2,
  gamma0 = 1,
  min.eps = 1e-07,
  maxit = 500
)
}
\description{
Newton-Raphson Optimization
}
