% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_removeNegative.R
\name{removeNegativeCount}
\alias{removeNegativeCount}
\title{' Remove Negative Counts
Removes the row that contains -1 in a specified column}
\usage{
removeNegativeCount(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{the name of the column}
}
\value{
returns the data frame with the rows containing -1 removed
}
\description{
' Remove Negative Counts
Removes the row that contains -1 in a specified column
}
