% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_replace0.R
\name{replace0}
\alias{replace0}
\title{Replace 0
Replace0 takes a specified column from the data frame and replaces cells that ahve the value 0 with NA}
\usage{
replace0(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{a column from the data frame}
}
\value{
returns the data frame with the 0's in the specified column changed
}
\description{
Replace 0
Replace0 takes a specified column from the data frame and replaces cells that ahve the value 0 with NA
}
