% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sepMonth.R
\name{separateCodedMonth}
\alias{separateCodedMonth}
\title{'Seperate Coded Month
Seperates a column containing coded year and coded month seperated by "-" into two columns and removes the input column}
\usage{
separateCodedMonth(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{the variable containing coded month and coded year}
}
\value{
returns the data frame with two new columns named "codedYear" and "codedMonth" and the input column removed
}
\description{
'Seperate Coded Month
Seperates a column containing coded year and coded month seperated by "-" into two columns and removes the input column
}
