% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sepDate.R
\name{separateDateOccur}
\alias{separateDateOccur}
\title{'Seperate Date Occur
Seperates a column containing month, day, year, and time into 4 different columns and removes the input column.}
\usage{
separateDateOccur(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{a column containing month, day, year, and time seperated by "|"}
}
\value{
returns 4 different columns containing, month, day, year, and time
}
\description{
'Seperate Date Occur
Seperates a column containing month, day, year, and time into 4 different columns and removes the input column.
}
