% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_sepIntersection.R
\name{separateIntersection}
\alias{separateIntersection}
\title{'Seperate Interection
Takes a column that contains street addresses and seperates the addresses into two columns}
\usage{
separateIntersection(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{the column containing street adresses}
}
\value{
returns two columns containing street addresses and removes the input column
}
\description{
'Seperate Interection
Takes a column that contains street addresses and seperates the addresses into two columns
}
