% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_crime.R
\name{cs_crime}
\alias{cs_crime}
\title{Identify Crimes}
\usage{
cs_crime(.data, var, newVar, crime)
}
\arguments{
\item{.data}{A tbl}

\item{var}{Name of variable with 5 or 6 digit crime codes}

\item{newVar}{Name of output variable to be created with logical data}

\item{crime}{A string describing the crime type to be identified}
}
\value{
A tibble with a logical vector that is \code{TRUE} if the given crime matches
    the category given in the function.
}
\description{
\code{cs_crime} can be used to easily identify
    crimes based on a specific single UCR categories or common groupings.
}
\details{
The categories used here are derived from the U.S. Federal
    Bureau of Investigation's Uniform Crime Reporting codes.
}
