% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_crime.R
\name{cs_crime}
\alias{cs_crime}
\title{Identify Violent Crimes}
\usage{
cs_crime(.data, var, crime)
}
\arguments{
\item{.data}{A tbl}

\item{var}{Name of variable with 5 or 6 digit crime codes}

\item{crime}{A string describing the crime type to be identified}
}
\value{
A logical vector
}
\description{
Violent crimes, as defined by the U.S. Federal Bureau of
    Investigation, are murder, rape, aggrevated assault, and burglary.
    \code{cs_crime_violent} can be used to identify specific single
    UCR categories or common groupings for purposes of variable
    creation or subsettings.
}
