% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{cs_crime_cat}
\alias{cs_crime_cat}
\title{Categorize Crime}
\usage{
cs_crime_cat(.data, var, newVar, output)
}
\arguments{
\item{.data}{A tbl}

\item{var}{Name of variable with 5 or 6 digit crime codes}

\item{newVar}{Name of output variable to be created with simplified categories}

\item{output}{Type of output - either string, factor, or numeric}
}
\value{
A copy of the data frame with the new output variable appended to it.
}
\description{
The SLMPD data contains 5 or 6 digit codes to refer to
    specific categories of crime. \code{cs_crime_cat} transforms these
    into either string, factor, or simplified numeric categories
    like "murder" or "aggrevated assault".
}
\details{
The categories used here are derived from the U.S. Federal
    Bureau of Investigation's Uniform Crime Reporting codes.
}
\examples{
testData <- january2018
cs_crime_cat(testData,Crime,SimpleCrime,output = "numeric")

}
