% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_filter_count.R, R/filter.R
\name{cs_filter_count}
\alias{cs_filter_count}
\title{Remove Negative Counts}
\usage{
cs_filter_count(.data, var)

cs_filter_count(.data, var)
}
\arguments{
\item{.data}{A tbl}

\item{var}{the name of the column}

\item{.data}{A tbl}

\item{var}{the name of the column}
}
\value{
returns the data frame with the rows containing -1 removed

returns the data frame with the rows containing -1 removed
}
\description{
Removes the row that contains -1 in a specified column

Removes the row that contains -1 in a specified column
}
\details{
Crimes with a count of -1 are crimes that were determined to be unfounded

Crimes with a count of -1 are crimes that were determined to be unfounded
}
\examples{
testData <- january2018
testData <- cs_filter_count(testData,Count)

testData <- january2018
testData <- cs_filter_count(testData,Count)

}
