% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_validate_year.R
\name{cs_identifyYear}
\alias{cs_identifyYear}
\title{Extract Year of a Given Year List Object Item}
\usage{
cs_identifyYear(monthItem)
}
\arguments{
\item{monthItem}{A single month from a year list object}
}
\description{
This uses the value of the first observation's coded month as the basis for
    identifying which month the data are from.
}
\keyword{internal}
