% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_missing_xy.R
\name{cs_missing_xy}
\alias{cs_missing_xy}
\title{Missing Coordinates}
\usage{
cs_missing_xy(.data, x, y)
}
\arguments{
\item{.data}{a data frame}

\item{x}{the column from the data frame containing the x coordinates}

\item{y}{the column from the data frame containg the y coordinates}
}
\value{
returns a logical vector that displays TRUE when the coordinates are
    missing and FALSE if the coordinates are not missing
}
\description{
\code{cs_missingCoords} compares X and Y coordinates and adds a
    TRUE/FALSE column that states which rows are missing coordinate
    data.
}
