% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_parse_date.R
\name{cs_parse_date}
\alias{cs_parse_date}
\title{Seperate Date Occur}
\usage{
cs_parse_date(.data, var, dateVar, timeVar, keepDateTime = TRUE)
}
\arguments{
\item{.data}{a data frame}

\item{var}{a column containing month, day, year, and time seperated by "/"}

\item{dateVar}{Name of variable containing date data}

\item{timeVar}{Name of variable containing time data}

\item{tz}{String name of timezone, defaults to system's timezone}

\item{keepDateTime}{A logical scalar. Keep an intermediate dateTime variable if \code{TRUE}.}
}
\value{
Appends two columns to the data frame. One is the time data and the other is the date data
}
\description{
Creates two columns. One contains month, day, and year and the other contains hour, and minute.
}
\examples{
testData <- january2018
testData <- cs_parse_date(testData, var = DateOccur, dateVar = Date, timeVar = Time)

}
