% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_parse_date.R
\name{cs_parse_date}
\alias{cs_parse_date}
\title{Seperate Date Occur}
\usage{
cs_parse_date(.data, variable)
}
\arguments{
\item{.data}{a data frame}

\item{variable}{a column containing month, day, year, and time seperated by "/"}
}
\value{
returns 4 different columns containing, month, day, year, and time
}
\description{
Seperates a column containing month, day, year, and time into 4 different
    columns and removes the input column.
}
