% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_parse_month.R
\name{cs_parse_month}
\alias{cs_parse_month}
\title{Seperate Coded Month}
\usage{
cs_parse_month(.data,var,newYear,newMonth)
}
\arguments{
\item{.data}{a data frame}

\item{var}{the variable containing coded month and coded year}

\item{newYear}{the name of the column to contain the year data}

\item{newMonth}{the name of the column to contain month data}
}
\value{
returns the data frame with two new columns named "codedYear" and "codedMonth" and the input column removed
}
\description{
Separates a column containing coded year and coded month
    separated by "-" into two columns and removes the input column
}
\examples{
testData <- january2018
testData <- cs_parse_month(testData,CodedMonth,Year,Month)


}
