% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_project.R
\name{cs_project}
\alias{cs_project}
\title{Project Data}
\usage{
cs_project(.data, area, varX, varY, crs)
}
\arguments{
\item{.data}{A tibble with crime data}

\item{area}{One of either "city" or "county"}

\item{varX}{Name of column containing x coordinate data}

\item{varY}{Name of column containing y coordinate data}

\item{crs}{integer with the EPSG code, or character with proj4string representing the
coordinate reference system}
}
\description{
\code{cs_project} converts STLMPD or St. Louis County data into a simple
    features object. For STLMPD data, this is done using the \code{XCoord} and
    \code{YCoord} columns. For St. Louis county data, this is done using the
    \code{X} and \code{Y} columns.
}
