% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{cs_replace0}
\alias{cs_replace0}
\title{Replace Coordinates with 0 Value with NA}
\usage{
cs_replace0(.data, var)
}
\arguments{
\item{.data}{A tibble or data frame}

\item{var}{Name of column containing coordinate data}
}
\value{
A tibble or data frame with the coordinate column updated.
}
\description{
This function a specified column from the data frame and
    replaces cells that have the value 0 with NA
}
\examples{
# load example data
testData <- january2018

# replace 0s in the x and y coordinate variables
testData <- cs_replace0(testData, var = XCoord)
testData <- cs_replace0(testData, var = YCoord)

}
