% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/january2018.R
\docType{data}
\name{january2018}
\alias{january2018}
\title{Crimes in St. Louis, January 2018}
\format{A tibble with 3825 rows and 20 variables:
\describe{
  \item{Compliant}{complaint record number}
  \item{CodedMonth}{year and month crime reported}
  \item{DateOccur}{date and time of crime}
  \item{FlagCrime}{}
  }}
\source{
\href{http://www.slmpd.org}{St. Louis Metropolitan Police Department}
}
\usage{
data(january2018)
}
\description{
A data set containing all reported crimes in St. Louis, Missouri during January 2018.
}
\examples{
str(january2018)
head(january2018)

}
\keyword{datasets}
